+++ src/Parser/Check/Dom/Text.php
@@ -33,6 +33,8 @@ class Text extends AbstractDomChecker
     {
         return ($this->node->parent()->tag != 'script'
             && $this->node->parent()->tag != 'style'
+            && $this->node->parent()->tag != 'noscript'
+            && $this->node->parent()->tag != 'code'
             && !is_numeric(TextUtil::fullTrim($this->node->outertext))
             && !preg_match('/^\d+%$/', TextUtil::fullTrim($this->node->outertext))
             && strpos($this->node->outertext, '[vc_') === false