+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.5.8] - 2018-06-26
+### Changed
+- Simplify code and fix all ignore nodes related problems
+### Fixed
+- Forgot noscript & code in Text checker
+
 ## [0.5.7] - 2018-06-12
 ### Changed
 - Client.HttpClient: Using new header system (Weglot-Context)